Remove PodSecurityPolicy admission plugin
