+++ pkg/kubelet/sysctl/allowlist_test.go
@@ -18,8 +18,6 @@ package sysctl
 
 import (
 	"testing"
-
-	"k8s.io/kubernetes/pkg/security/podsecuritypolicy/sysctl"
 )
 
 func TestNewAllowlist(t *testing.T) {
@@ -37,7 +35,7 @@ func TestNewAllowlist(t *testing.T) {
 		{sysctls: []string{"net.*/foo"}, err: true},
 		{sysctls: []string{"foo"}, err: true},
 	} {
-		_, err := NewAllowlist(append(sysctl.SafeSysctlAllowlist(), test.sysctls...))
+		_, err := NewAllowlist(append(SafeSysctlAllowlist(), test.sysctls...))
 		if test.err && err == nil {
 			t.Errorf("expected an error creating a allowlist for %v", test.sysctls)
 		} else if !test.err && err != nil {
@@ -69,7 +67,7 @@ func TestAllowlist(t *testing.T) {
 		{sysctl: "kernel.sem", hostIPC: true},
 	}
 
-	w, err := NewAllowlist(append(sysctl.SafeSysctlAllowlist(), "kernel.msg*", "kernel.sem"))
+	w, err := NewAllowlist(append(SafeSysctlAllowlist(), "kernel.msg*", "kernel.sem"))
 	if err != nil {
 		t.Fatalf("failed to create allowlist: %v", err)
 	}
