--- pkg/security/podsecuritypolicy/OWNERS
@@ -1,8 +1,0 @@
-# See the OWNERS docs at https://go.k8s.io/owners
-
-approvers:
-  - sig-auth-policy-approvers
-reviewers:
-  - sig-auth-policy-reviewers
-labels:
-  - sig/auth
