--- pkg/security/podsecuritypolicy/user/runasany_test.go
@@ -1,59 +1,0 @@
-/*
-Copyright 2016 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package user
-
-import (
-	"testing"
-
-	policy "k8s.io/api/policy/v1beta1"
-)
-
-func TestRunAsAnyOptions(t *testing.T) {
-	_, err := NewRunAsAny(nil)
-	if err != nil {
-		t.Fatalf("unexpected error initializing NewRunAsAny %v", err)
-	}
-	_, err = NewRunAsAny(&policy.RunAsUserStrategyOptions{})
-	if err != nil {
-		t.Errorf("unexpected error initializing NewRunAsAny %v", err)
-	}
-}
-
-func TestRunAsAnyGenerate(t *testing.T) {
-	s, err := NewRunAsAny(&policy.RunAsUserStrategyOptions{})
-	if err != nil {
-		t.Fatalf("unexpected error initializing NewRunAsAny %v", err)
-	}
-	uid, err := s.Generate(nil, nil)
-	if uid != nil {
-		t.Errorf("expected nil uid but got %d", *uid)
-	}
-	if err != nil {
-		t.Errorf("unexpected error generating uid %v", err)
-	}
-}
-
-func TestRunAsAnyValidate(t *testing.T) {
-	s, err := NewRunAsAny(&policy.RunAsUserStrategyOptions{})
-	if err != nil {
-		t.Fatalf("unexpected error initializing NewRunAsAny %v", err)
-	}
-	errs := s.Validate(nil, nil, nil, nil, nil)
-	if len(errs) != 0 {
-		t.Errorf("unexpected errors validating with ")
-	}
-}
