--- pkg/security/podsecuritypolicy/user/types.go
@@ -1,31 +1,0 @@
-/*
-Copyright 2016 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package user
-
-import (
-	"k8s.io/apimachinery/pkg/util/validation/field"
-	api "k8s.io/kubernetes/pkg/apis/core"
-)
-
-// RunAsUserStrategy defines the interface for all uid constraint strategies.
-type RunAsUserStrategy interface {
-	// Generate creates the uid based on policy rules.
-	Generate(pod *api.Pod, container *api.Container) (*int64, error)
-	// Validate ensures that the specified values fall within the range of the strategy.
-	// scPath is the field path to the container's security context
-	Validate(scPath *field.Path, pod *api.Pod, container *api.Container, runAsNonRoot *bool, runAsUser *int64) field.ErrorList
-}
